requirements.txt
opencv-python==4.2.0.32
numpy==1.19.1
Flask==1.0.2
Flask-Cors==3.0.7
pillow==7.1.0
paddlepaddle==1.8.5
gunicorn==20.0.4
onnxruntime==1.4.0
pyclipper==1.2.0
shapely
moviepy==1.0.3
pydub==0.25.1
opencv-contrib-python==4.5.2.54
ffmpy3==0.2.4
pycorrector
kenlm
Scikit-Image
sklearn
imagehash
airtest
APScheduler==3.7.0